Session storing in its own file to prevent circular dependencies
